Pause field now editable, changes will be accounted for
When updating a shift, the pause duration will be shown to the user.
Also when the value for the pause duration is changed, the total work
time will be recalculated.
This DOES NOT check for negative total work time values, as we don't
want to hassle the user. If the user doesn't care, then he'll have
negative work time values in his exported sheet.