Update test case
That test does not produce consistent data. We are trying to compare shifts and
make sure that they are sorted accordingly. The problem here is that we can also
get shifts with the same finished datetimes. In this case, we will skip the
assertion lazily.

Not a good solution, but it does the job for now and we will probably rewrite
that functionality soon or delete it completely.